pyproject.toml
[tool.poetry]
name = "streamlit-sandbox"
version = "0.1.0"
description = ""
authors = ["Anastasiia Kulakova <anastasiia.kulakova@jedlix.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "1.9.0"
protobuf = "3.19.5"
openai = "0.26.4"
python-decouple = "^3.7"
google-cloud-vision = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
